+++ TraderBot.EmailListener/HostedServices/EmailListenerStatus.cs
@@ -1,0 +1,9 @@
+namespace TraderBot.EmailListener.HostedServices;
+
+public enum EmailListenerStatus
+{
+    New,
+    Connected,
+    Wait,
+    Error
+}