+++ TraderBot.RavenDb/RavenDbOptions.cs
@@ -2,6 +2,6 @@ namespace TraderBot.RavenDb;
 
 public class RavenDbOptions
 {
-    public IEnumerable<string> Urls { get; set; } = new[] { "http://127.0.0.1:8080" };
-    public string DatabaseName { get; set; } = "TraderBor";
+    public IEnumerable<string> Urls { get; set; } = Array.Empty<string>();
+    public string DatabaseName { get; set; } = "TraderBot";
 }