+++ CHANGELOG.md
@@ -8,7 +8,9 @@ Use [Keep a Changelog](https://keepachangelog.com/ru/1.0.0/) style for any chang
 
 ### Added
 
-+ Control trade copy by allow list. Only allowed user trades will be copied
++ OrderController: Control trade copy by allow list. Only allowed user trades will be copied
++ EmailListener: Support multiple mail box accounts by `MailBoxStore`. Each mail box processed by `EmailListenerStateMachine`
++ EmailListener: Communication between `EmailListenerJob` and `CollectProcessCommand` processed with `System.Threading.Channels` by `EmailChannel` 
 
 ### Changed